Worked a lot on the second system calculations and saving the results

+++ src/main/java/com/example/bettingGame/core/dto/UserScoreDto.java
@@ -8,5 +8,5 @@ public class UserScoreDto {
 
     private Long userId;
     private String username;
-    private Integer numberOfPoints;
+    private Double numberOfPoints;
 }
